pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "chatbot_conversation"
description = "A chatbot conversation system"
version = "0.1.0"
authors = [{name = "Brian Sentance", email = "brian@sentance.tech"}]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "anthropic==0.42.0",
    "google-generativeai==0.8.3",
    "ollama==0.4.5",
    "openai==1.59.6",
    "pydantic==2.10.5",
    "python-dotenv==1.0.1",
    "rich==13.9.4",
    "tenacity==8.5.0"
]

[project.scripts]
chatbot_conversation = "chatbot_conversation.main:main"

[tool.hatch.build.targets.wheel]
packages = ["src/chatbot_conversation"]

[tool.hatch.metadata]
root = "src"

[tool.hatch.envs.test]
dependencies = [
    "black=24.10.0",
    "hatch=1.14.0",
    "hatchling=1.25.0",
    "isort=5.13.2",
    "mypy=1.11.2",
    "mypy-extensions=1.0.0",
    "pipreqs=0.5.0",
    "pylint=3.2.7",
    "pytest=7.4.4",
    "pytest-cov=6.0.0",
    "tomlcheck=0.2.3"
]

[tool.isort]
profile = "black"
multi_line_output = 3

[tool.black]
line-length = 88
target-version = ['py38']

[tool.mypy]
mypy_path = ["src"]
python_version = "3.10"
strict = true
ignore_missing_imports = false
explicit_package_bases = true

[tool.pytest.ini_options]
pythonpath = [".", "src"]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v -s"
filterwarnings = [
    "ignore::DeprecationWarning",
    "ignore::UserWarning",
]
